Ensure dot commands don't have cribl prepended
