+++ src/datasource.ts
@@ -12,6 +12,7 @@ import {
 import { getBackendSrv } from '@grafana/runtime';
 import { lastValueFrom } from 'rxjs';
 import { CriblQuery, CriblDataSourceOptions } from './types';
+import { prependCriblOperator } from 'utils';
 
 const MAX_RESULTS = 10000; // same as what the actual Cribl UI imposes
 const QUERY_PAGE_SIZE = 1000;
@@ -350,28 +351,4 @@ function parseJwtExp(token: string): number {
       return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
   }).join(''));
   return JSON.parse(jsonPayload).exp;
-}
-
-/**
- * Cribl Search backend requires a fully-formed query including the "cribl" operator.
- * Here we're applying best effort to auto-prepending the "cribl" operator where it needs to be.
- * This is far from 100% foolproof, but covers 99% of the happy paths.
- * @param query the query as the user entered it
- * @returns the query with "cribl" operator prepended as needed, if we were able to
- */
-export function prependCriblOperator(query: string): string {
-  // This tries to capture the first "word" of the root query statement, along with any preceding
-  // "set" or "let" statements, and everything that comes after the first word.  NOTE: This doesn't
-  // touch any "let" statements.  We might enhance that in the future, but for now, users need to
-  // put their own "cribl" operator in those stage statements.
-  const firstWordRegex = /^((?:\s*(?:set|let)\s+[^;]+;)*\s*)((\w+|['"*]).*)$/;
-  const matches = query.trim().match(firstWordRegex);
-  if (matches != null) {
-    const firstWord = matches[3];
-    // We recognize certain operators that don't need "cribl" prepended
-    if (!['cribl', 'externaldata', 'find', 'print', 'search'].includes(firstWord)) {
-      return `${matches[1]}cribl ${matches[2]}`;
-    }
-  }
-  return query;
 }
